nsp-dbrd/nsp-dbrd/backend/requirements.txt
Flask==2.0.3
psycopg2-binary==2.9.1
pandas==1.3.3
numpy==1.21.2
scipy==1.7.1
logging==0.5.1.2
```